streamline fns and fix prompts
